ADFA-5240: Move the compression-dictionary loader into :common
WebServer decided on its own whether a database's brotli Content rows
carry the shared dictionary. PluginDocumentationManager, in another
module, is about to need the identical decision when it writes those
rows -- and a writer that disagrees with the reader produces content
nothing can decode.

Move loadCompressionDictionary and toDirectByteBuffer to :common,
next to the DatabaseVersionResolver they gate on, so both sides read
the one implementation.

Pure move: the version gate, the CompressionDictionary checks and the
throw-vs-null contract are unchanged, as is WebServer's retry.
